+++ test/functional/crossFunctional.mocha.js
@@ -482,6 +482,24 @@ describe("Cross Functional", function() {
         .limit(20)
     }));
 
+    it('works with STRING split .fallback() no match', equalityTest({
+      executorNames: ['druid', 'mysql', 'postgres'],
+      expression: $('wiki').split('$cityName.fallback("NoCity")', 'CityFallback')
+        .apply('TotalEdits', '$wiki.sum($count)')
+        .apply('TotalAdded', '$wiki.sum($added)')
+        .sort('$TotalAdded', 'descending')
+        .limit(20)
+    }));
+
+    it('works with STRING split .fallback() with match', equalityTest({
+      executorNames: ['druid', 'mysql', 'postgres'],
+      expression: $('wiki').split('$cityName.fallback("Bucharest")', 'CityFallback')
+        .apply('TotalEdits', '$wiki.sum($count)')
+        .apply('TotalAdded', '$wiki.sum($added)')
+        .sort('$TotalAdded', 'descending')
+        .limit(20)
+    }));
+
     it('works with NUMBER split (numberBucket) (sort on split)', equalityTest({
       executorNames: ['druid', 'mysql', 'postgres'],
       expression: $('wiki').split($("commentLength").numberBucket(10), 'CommentLength10')